requirements.txt
pillow
pytesseract
opencv-python-headless
numpy
requests
freetype-py
pkg_resources
qrcode[pil]
cairosvg
wand
easyocr
